Feat: Add recent Ubuntu 18.04 and 20.04 images
In light of the recent expiration of the Let's Encrypt cross-signing CA
building new images based on the currently provided Ubuntu images may
have issues when used in conjuction with packer as they require a reboot
during the setup.

Adding in latest point releases of both LTS distros should resolve this
issue.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>